start moving tests into more appropriate folders
