+++ test/__tests/methods/animate.js
