+++ test/__tests/methods/merge.js
