+++ test/__tests/methods/toggle.js
